refactor: restructure library files
