+++ packages/react-three-rapier/src/utils/interaction-groups.ts
