+++ packages/react-three-rapier/src/utils/shared-objects.ts
