+++ packages/react-three-rapier/src/utils/utils-rigidbody.ts
@@ -1,8 +1,12 @@
 import { RigidBody, RigidBodyDesc } from "@dimforge/rapier3d-compat";
 import React, { MutableRefObject, useEffect, useMemo } from "react";
 import { Matrix4, Object3D, Vector3 } from "three";
-import { Boolean3Array, RigidBodyProps, Vector3Array } from ".";
-import { EventMap, RigidBodyState, RigidBodyStateMap } from "./Physics";
+import { Boolean3Array, RigidBodyProps, Vector3Array } from "..";
+import {
+  EventMap,
+  RigidBodyState,
+  RigidBodyStateMap
+} from "../components/Physics";
 import {
   _matrix4,
   _position,
